COMMERCE-1061 Create interface because if we want to inject this via Context Provider into field variable we need to use this way
